Revert extracting DB helper functions to DbHelper.php
