+++ protected/humhub/tests/codeception/_support/HumHubHelperTrait.php
@@ -6,13 +6,15 @@
  * @license   https://www.humhub.com/licences
  */
 
+/**
+ * @noinspection PhpIllegalPsrClassPathInspection
+ */
+
 namespace tests\codeception\_support;
 
 use Codeception\Exception\ModuleException;
-use Codeception\Module;
 use Codeception\Module\Yii2;
 use humhub\components\behaviors\PolymorphicRelation;
-use humhub\helpers\DbHelper;
 use humhub\models\UrlOembed;
 use humhub\modules\activity\models\Activity;
 use humhub\modules\content\widgets\richtext\converter\RichTextToHtmlConverter;
@@ -27,7 +29,9 @@ use Yii;
 use yii\base\ErrorException;
 use yii\base\InvalidConfigException;
 use yii\db\ActiveRecord;
+use yii\db\Command;
 use yii\db\ExpressionInterface;
+use yii\db\Query;
 use yii\helpers\FileHelper;
 use yii\log\Dispatcher;
 
@@ -431,7 +435,7 @@ trait HumHubHelperTrait
      */
     public static function assertRecordCount(?int $expected, $tables, $condition = null, ?array $params = [], string $message = ''): void
     {
-        $count = DbHelper::dbCount($tables, $condition, $params ?? []);
+        $count = static::dbCount($tables, $condition, $params ?? []);
 
         if ($expected === null) {
             static::assertGreaterThan(0, $count, $message);
@@ -495,7 +499,7 @@ trait HumHubHelperTrait
      */
     public static function assertRecordValue($expected, string $column, $tables, $condition = null, ?array $params = [], string $message = ''): void
     {
-        $value = DbHelper::dbQuery($tables, $condition, $params, 1)->select($column)->scalar();
+        $value = self::dbQuery($tables, $condition, $params, 1)->select($column)->scalar();
         static::assertEquals($expected, $value, $message);
     }
 
@@ -672,5 +676,106 @@ trait HumHubHelperTrait
     public static function logFilterMessageTexts($levels = null, ?array $categories = null, ?array $exceptCategories = null): array
     {
         return array_column(static::logFilterMessages($levels, $categories, $exceptCategories), 0);
+    }
+
+    /**
+     * @see \yii\db\Connection::createCommand()
+     * @since 1.15
+     */
+    public static function dbCommand($sql = null, $params = []): Command
+    {
+        return Yii::$app->getDb()->createCommand($sql, $params);
+    }
+
+    /**
+     * @param Command $cmd
+     * @param bool $execute
+     *
+     * @return Command
+     * @throws \yii\db\Exception
+     */
+    protected static function dbCommandExecute(Command $cmd, bool $execute = true): Command
+    {
+        if ($execute) {
+            $cmd->execute();
+        }
+
+        return $cmd;
+    }
+
+    /**
+     * @see Query
+     * @since 1.15
+     */
+    public static function dbQuery($tables, $condition, $params = [], $limit = 10): Query
+    {
+        return (new Query())
+            ->from($tables)
+            ->where($condition, $params)
+            ->limit($limit);
+    }
+
+    /**
+     * @see Command::insert
+     * @since 1.15
+     */
+    public static function dbInsert($table, $columns, bool $execute = true): Command
+    {
+        return static::dbCommandExecute(static::dbCommand()->insert($table, $columns), $execute);
+    }
+
+    /**
+     * @see Command::update
+     * @since 1.15
+     */
+    public static function dbUpdate($table, $columns, $condition = '', $params = [], bool $execute = true): Command
+    {
+        return static::dbCommandExecute(static::dbCommand()->update($table, $columns, $condition, $params), $execute);
+    }
+
+    /**
+     * @see Command::upsert
+     * @since 1.15
+     */
+    public static function dbUpsert($table, $insertColumns, $updateColumns = true, $params = [], bool $execute = true): Command
+    {
+        return static::dbCommandExecute(
+            static::dbCommand()->upsert($table, $insertColumns, $updateColumns, $params),
+            $execute
+        );
+    }
+
+    /**
+     * @see Command::delete()
+     * @since 1.15
+     */
+    public static function dbDelete($table, $condition = '', $params = [], bool $execute = true): Command
+    {
+        return static::dbCommandExecute(static::dbCommand()->delete($table, $condition, $params), $execute);
+    }
+
+    /**
+     * @see Query::select
+     * @see Query::from
+     * @see Query::where
+     * @see \yii\db\QueryTrait::limit()
+     * @since 1.15
+     */
+    public static function dbSelect($tables, $columns, $condition = '', $params = [], $limit = 10, $selectOption = null): array
+    {
+        return static::dbQuery($tables, $condition, $params, $limit)
+            ->select($columns, $selectOption)
+            ->all();
+    }
+
+    /**
+     * @see Command::delete()
+     * @since 1.15
+     */
+    public static function dbCount($tables, $condition = '', $params = [])
+    {
+        return static::dbQuery($tables, $condition, $params)
+            ->select("count(*)")
+            ->scalar();
     }
 }
